chore: Update useful commands and proguard rules
This commit updates the useful commands section in the README.md file.. Additionally, the proguard-rules.pro file is updated to retain generic signatures of TypeToken and its subclasses with R8 version 3.0 and higher, and to keep the model classes used by Gson. These changes improve the documentation and configuration of the project.